+++ src/py_bugger/py_bugger.py
@@ -29,6 +29,9 @@ def main(exception_type, target_dir, target_file, num_bugs):
     elif exception_type == "AttributeError":
         new_bugs_made = buggers.attribute_error_bugger(py_files, num_bugs)
         bugs_added += new_bugs_made
+    elif exception_type == "IndentationError":
+        new_bugs_made = buggers.indentation_error_bugger(py_files, num_bugs)
+        bugs_added += new_bugs_made
 
     # Show a final success/fail message.
     if bugs_added == num_bugs: