+++ tests/sample_code/sample_scripts/simple_indent.py
@@ -1,0 +1,2 @@
+for num in [1, 2, 3]:
+    print(num)